backend/requirements.txt
fastapi
uvicorn
transformers
torch
groq
python-dotenv
google-genai